update(middleware):
adds middleware for routes

+++ src/controllers/notes.ts
@@ -2,10 +2,13 @@ import { RequestHandler } from "express";
 import NoteModel from "../models/note";
 import createHttpError from "http-errors";
 import mongoose from "mongoose";
+import { assertIsDefined } from "../util/assertIsDefined";
 
 export const getNotes: RequestHandler = async (req, res, next) => {
+  const authenticatedUserId = req.session.userId;
   try {
-    const notes = await NoteModel.find().exec();
+    assertIsDefined(authenticatedUserId);
+    const notes = await NoteModel.find({ userId: authenticatedUserId }).exec();
     res.status(200).json(notes);
   } catch (err) {
     next(err);
@@ -14,13 +17,20 @@ export const getNotes: RequestHandler = async (req, res, next) => {
 
 export const getNote: RequestHandler = async (req, res, next) => {
   const noteId = req.params.noteId;
+  const authenticatedUserId = req.session.userId;
+
   try {
+    assertIsDefined(authenticatedUserId);
+
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(400, "Invalid note id");
     }
     const note = await NoteModel.findById(noteId).exec();
     if (!note) {
       throw createHttpError(404, "Note not found");
+    }
+    if (!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, "You can not access this note!");
     }
     res.status(200).json(note);
   } catch (error) {
@@ -41,11 +51,16 @@ export const createNote: RequestHandler<
 > = async (req, res, next) => {
   const title = req.body.title;
   const text = req.body.text;
+  const authenticatedUserId = req.session.userId;
+
   try {
+    assertIsDefined(authenticatedUserId);
+
     if (!title) {
       throw createHttpError(400, "Note must have a title");
     }
     const newNote = await NoteModel.create({
+      userId: authenticatedUserId,
       title: title,
       text: text,
     });
@@ -71,18 +86,26 @@ export const updateNote: RequestHandler<
   const noteId = req.params.noteId;
   const newTitle = req.body.title;
   const newText = req.body.text;
+  const authenticatedUserId = req.session.userId;
+
   try {
+    assertIsDefined(authenticatedUserId);
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(400, "Invalid note id");
     }
+    if (!newTitle) {
+      throw createHttpError(400, "Note must have a title");
+    }
+
     const note = await NoteModel.findById(noteId).exec();
 
     if (!note) {
       throw createHttpError(404, "Note not found");
     }
-    if (!newTitle) {
-      throw createHttpError(400, "Note must have a title");
+    if (!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, "You can not access this note!");
     }
+
     note.title = newTitle;
     note.text = newText;
     const updatedNote = await note.save();
@@ -94,8 +117,11 @@ export const updateNote: RequestHandler<
 
 export const deleteNote: RequestHandler = async (req, res, next) => {
   const noteId = req.params.noteId;
+  const authenticatedUserId = req.session.userId;
 
   try {
+    assertIsDefined(authenticatedUserId);
+
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(400, "Invalid note id");
     }
@@ -103,6 +129,9 @@ export const deleteNote: RequestHandler = async (req, res, next) => {
 
     if (!note) {
       throw createHttpError(404, "Note not found");
+    }
+    if (!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, "You can not access this note!");
     }
     await note.deleteOne();
     res.sendStatus(204);
